chore(medicines): move medicine query to medicine.ts
specifically services/medicine.ts
